split AllTaskProcessedDataset, assert_no_double_bos, collate_fn to different files
