+++ nemo_rl/data/collate_fn.py
@@ -11,18 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Optional, Union
+from typing import Any, Union
 
 import torch
-from datasets import Dataset
 from transformers import AutoProcessor, PreTrainedTokenizerBase
 
-from nemo_rl.data.interfaces import (
-    DatumSpec,
-    DPODatumSpec,
-    TaskDataProcessFnCallable,
-    TaskDataSpec,
-)
+from nemo_rl.data.interfaces import DatumSpec, DPODatumSpec
 from nemo_rl.data.llm_message_utils import (
     add_loss_mask_to_message_log,
     batched_message_log_to_flat_message,
@@ -30,105 +24,6 @@ from nemo_rl.data.llm_message_utils import (
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
 
 TokenizerType = Union[PreTrainedTokenizerBase, AutoProcessor]
-
-
-# TODO @sahilj handle too-long prompts and masking them out throughout the whole process and renormalizing on loss
-class AllTaskProcessedDataset:
-    """Dataset for processing single or multi-task data with task-specific tokenization and processing.
-
-    Args:
-        dataset: Input dataset containing raw data
-        tokenizer: Tokenizer for text processing
-        default_task_data_spec: Default task processing specifications.
-            In the case of single-task, this is the spec used for processing all entries.
-            In the case of multi-task, any values not specified in the task-specific specs will be taken from the default spec.
-        task_data_processors: Either a single TaskDataProcessFnCallable for single-task,
-            or a dict mapping task names to (TaskDataSpec, TaskDataProcessFnCallable) for multi-task
-        max_seq_length: Maximum sequence length for tokenized outputs
-    """
-
-    def __init__(
-        self,
-        dataset: Dataset | Any,
-        tokenizer: TokenizerType,
-        default_task_data_spec: TaskDataSpec,
-        task_data_processors: (
-            dict[str, tuple[TaskDataSpec, TaskDataProcessFnCallable]]
-            | TaskDataProcessFnCallable
-        ),
-        max_seq_length: Optional[int] = None,
-    ):
-        self.dataset = dataset
-        self.tokenizer = tokenizer
-        self.default_task_data_spec = default_task_data_spec
-        self.task_data_processors = task_data_processors
-        self.max_seq_length = max_seq_length
-        self._bos_checked = False
-
-        if isinstance(task_data_processors, dict):
-            # apply defaults to all task data specs
-            for task_name, (
-                task_data_spec,
-                task_data_processor,
-            ) in task_data_processors.items():
-                task_data_spec.copy_defaults(self.default_task_data_spec)
-
-    def __len__(self) -> int:
-        return len(self.dataset)
-
-    def encode_single(
-        self, text: Union[str, list[str]]
-    ) -> tuple[list[int] | torch.Tensor, int]:
-        """Takes either a single string or a list of strings that represent multiple turns for the same conversation.
-
-        Returns a single (concatenated) list of tokenized ids and the length of the tokenized ids.
-        """
-        if isinstance(text, str):
-            text_ids = self.tokenizer.text_to_ids(text)
-            return text_ids, len(text_ids)
-        elif isinstance(text, list):
-            text_ids = [self.tokenizer.text_to_ids(t) for t in text]
-            return torch.cat(text_ids), sum(len(t) for t in text_ids)
-        else:
-            raise ValueError(
-                f"text must be a string or a list of strings, got {type(text)}"
-            )
-
-    def __getitem__(self, idx: int) -> DatumSpec:
-        """Return a single prompt."""
-        entry = self.dataset[idx]
-
-        if isinstance(self.task_data_processors, dict):
-            task_name = entry["task_name"]
-
-            assert task_name in self.task_data_processors, (
-                f"task processor not provided for {task_name}. Provided processors: {self.task_data_processors.keys()}"
-            )
-            task_data_spec, task_data_processor = self.task_data_processors[task_name]
-        else:
-            task_data_spec = self.default_task_data_spec
-            task_data_processor = self.task_data_processors
-
-        datum_spec = task_data_processor(
-            entry, task_data_spec, self.tokenizer, self.max_seq_length, idx
-        )
-
-        # Check the first processed item for BOS token assertion
-        if (
-            not self._bos_checked
-            and "message_log" in datum_spec
-            and datum_spec["message_log"]
-        ):
-            first_message = datum_spec["message_log"][0]
-            if "token_ids" in first_message:
-                token_ids = first_message["token_ids"]
-                assert isinstance(token_ids, torch.Tensor), (
-                    f"token_ids must be a torch.Tensor, got {type(token_ids)}"
-                )
-                assert_no_double_bos(token_ids, self.tokenizer)
-            self._bos_checked = True
-
-        return datum_spec
 
 
 def rl_collate_fn(data_batch: list[DatumSpec]) -> BatchedDataDict[Any]:
@@ -193,7 +88,7 @@ def eval_collate_fn(data_batch: list[DatumSpec]) -> BatchedDataDict[Any]:
     Examples:
     ```{doctest}
     >>> import torch
-    >>> from nemo_rl.data.datasets import eval_collate_fn
+    >>> from nemo_rl.data.collate_fn import eval_collate_fn
     >>> from nemo_rl.data.interfaces import DatumSpec
     >>> data_batch = [
     ...     DatumSpec(
@@ -300,24 +195,3 @@ def preference_collate_fn(
         data["token_mask"] = cat_and_padded["token_loss_mask"]
 
     return data
-
-
-def assert_no_double_bos(token_ids: torch.Tensor, tokenizer: TokenizerType) -> None:
-    """Assert that there are no double starting BOS tokens in the message.
-
-    Args:
-        token_ids: List of token IDs
-        tokenizer: Tokenizer
-    """
-    if tokenizer.bos_token_id is not None:
-        token_ids_list = token_ids.tolist()
-        if len(token_ids_list) > 1:
-            assert not (
-                token_ids_list[0] == tokenizer.bos_token_id
-                and token_ids_list[1] == tokenizer.bos_token_id
-            ), "Found double BOS token in the first two positions of the message."
-    else:
-        # `name_or_path` is not available for AutoProcessor, temp fix in get_tokenizer
-        print(
-            f"skip assert_start_single_bos since Tokenizer {tokenizer.name_or_path} has no BOS token"
-        )
