+++ nemo_rl/evals/eval.py
@@ -26,7 +26,8 @@ from transformers import AutoTokenizer
 
 from nemo_rl.algorithms.utils import set_seed
 from nemo_rl.data import MathDataConfig
-from nemo_rl.data.datasets import AllTaskProcessedDataset, eval_collate_fn
+from nemo_rl.data.collate_fn import eval_collate_fn
+from nemo_rl.data.datasets import AllTaskProcessedDataset
 from nemo_rl.data.llm_message_utils import get_keys_from_message_log
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
 from nemo_rl.distributed.virtual_cluster import ClusterConfig, RayVirtualCluster
