+++ tests/unit/data/test_datasets.py
@@ -16,7 +16,7 @@ from unittest.mock import MagicMock
 
 import torch
 
-from nemo_rl.data.datasets import preference_collate_fn
+from nemo_rl.data.collate_fn import preference_collate_fn
 from nemo_rl.data.interfaces import DatumSpec
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
